Module vmm_sys_util::aio

Safe wrapper over Linux AIO.

Structs

IoContext

Newtype for aio_context_t.

IoControlBlock

Wrapper over the iocb structure.

IoEvent

Wrapper over the io_event structure.

Constants

IOCB_CMD_FDSYNC

Synchronize a file’s in-core data to storage device.

IOCB_CMD_FSYNC

Synchronize a file’s in-core metadata and data to storage device.

IOCB_CMD_NOOP

Noop, this defined by never used by linux kernel.

IOCB_CMD_PREAD

Read from a file descriptor at a given offset.

IOCB_CMD_PREADV

Read from a file descriptor at a given offset into multiple buffers.

IOCB_CMD_PWRITE

Write to a file descriptor at a given offset.

IOCB_CMD_PWRITEV

Write to a file descriptor at a given offset from multiple buffers.

IOCB_FLAG_RESFD

Valid flags for the “aio_flags” member of the “struct iocb”. Set if the “aio_resfd” member of the “struct iocb” is valid.

MAX_REQUESTS

Maximum number of concurrent requests.