Module vmm_sys_util::errno

Structures, helpers, and type definitions for working with errno.

Structs

Error

Wrapper over errno.

Functions

errno_result

Returns the last errno as a Result that is always an error.

Type Definitions

Result

A specialized Result type for operations that can return errno.