Function vmm_sys_util::ioctl::ioctl_with_mut_ref

pub unsafe fn ioctl_with_mut_ref<F: AsRawFd, T>(
    fd: &F, 
    req: c_ulong, 
    arg: &mut T
) -> c_int

Run an ioctl with a mutable reference.

Arguments

fd: an open file descriptor corresponding to the device on which to call the ioctl.

req: a device-dependent request code.

arg: a mutable reference passed to ioctl.

Safety

The caller should ensure to pass a valid file descriptor and have the return value checked.