Trait vmm_sys_util::signal::Killable

pub unsafe trait Killable {
    fn pthread_handle(&self) -> pthread_t;

    fn kill(&self, num: c_int) -> Result<()> { ... }
}

Trait for threads that can be signalled via pthread_kill.

Note that this is only useful for signals between SIGRTMIN() and SIGRTMAX() because these are guaranteed to not be used by the C runtime.

Safety

This is marked unsafe because the implementation of this trait must guarantee that the returned pthread_t is valid and has a lifetime at least that of the trait object.

Required methods

fn pthread_handle(&self) -> pthread_t

Cast this killable thread as pthread_t.

Provided methods

fn kill(&self, num: c_int) -> Result<()>

Send a signal to this killable thread.

Arguments

num: specify the signal